Added documentation and reorganized components
