+++ src/pages/index.tsx
@@ -4,8 +4,8 @@ import { LunchHitchUser } from '../auth';
 import AuthSelector from '../common/auth_selector';
 import Navbar from '../common/components/navbar';
 
-import NoUserHomePage from './profile/noUser/NoUserHomePage';
-import UserHomePage from './profile/user/UserHomePage';
+import NoUserHomePage from './profile/noUser';
+import UserHomePage from './profile/user';
 
 export default function IndexPage() {
   return (
